Use new lsst.alert.stream.serialization module
