+++ test/testutils.py
@@ -19,11 +19,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 import io
-import os
-import json
 
 import astropy.time
-import fastavro
 
 from lsst.alert.packet import SchemaRegistry
